pyproject.toml
[project]
name = "annas-gif-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.8",
    "fuzzywuzzy>=0.18.0",
    "jinja2>=3.1.5",
    "moviepy==1.0.3",
    "nltk>=3.9.1",
    "openai-whisper>=20240930",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "sse-starlette>=1.6.5",
]
